PoseDetector/requirements.txt
streamlit
mediapipe
opencv-python
numpy
pillow